Line diagram api WIP
